+++ python/dolfin/common.py
@@ -120,81 +120,3 @@ def timed(task: str):
                 return func(*args, **kwargs)
         return wrapper
     return decorator
-
-
-if has_mpi4py:
-
-    class MPICommWrapper:
-        """DOLFIN is compiled without support for mpi4py. This object
-        can be passed into DOLFIN as an MPI communicator, but is not
-        an mpi4py comm.
-        """
-
-        def underlying_comm(self):
-            return cpp.MPICommWrapper.underlying_comm()
-
-
-class MPI:
-    """MPI parallel communication wrapper
-    """
-
-    comm_world = cpp.MPI.comm_world
-    comm_self = cpp.MPI.comm_self
-    comm_null = cpp.MPI.comm_null
-
-    @staticmethod
-    def init():
-        """Initilize MPI
-        """
-        cpp.MPI.init()
-
-    @staticmethod
-    def init_level(args: list, required_thread_level: int):
-        """Initialise MPI with command-line args and required level
-        of thread support.
-        """
-        cpp.MPI.init(args, required_thread_level)
-
-    @staticmethod
-    def responsible():
-        return cpp.MPI.responsible()
-
-    @staticmethod
-    def initialized():
-        return cpp.MPI.initialized()
-
-    @staticmethod
-    def finalized():
-        return cpp.MPI.finalized()
-
-    @staticmethod
-    def barrier(comm: MPICommWrapper):
-        return cpp.MPI.barrier(comm)
-
-    @staticmethod
-    def rank(comm: MPICommWrapper):
-        return cpp.MPI.rank(comm)
-
-    @staticmethod
-    def size(comm: MPICommWrapper):
-        return cpp.MPI.size(comm)
-
-    @staticmethod
-    def local_range(comm: MPICommWrapper, N: int):
-        return cpp.MPI.local_range(comm, N)
-
-    @staticmethod
-    def max(comm: MPICommWrapper, value: float):
-        return cpp.MPI.max(comm, value)
-
-    @staticmethod
-    def min(comm: MPICommWrapper, value: float):
-        return cpp.MPI.min(comm, value)
-
-    @staticmethod
-    def sum(comm: MPICommWrapper, value: float):
-        return cpp.MPI.sum(comm, value)
-
-    @staticmethod
-    def avg(comm: MPICommWrapper, value):
-        return cpp.MPI.avs(comm, value)
